Subject: Scraping Process
In this phase I completed scraping data from website. For scraping I
used selenium because website is dynamic and hide data so it is
difficult for Scrapy framework to scrap. I tried Scrpay but it returns
null values for same xpath which I used in selenium for extracting data
the reason is that the data is dynimacly changing.