Subject: Database Integeration
In this phase, I integerate MySQL database with python to store data.
During that process first it's take information for bulding database
connection after connecting it's check wheter the desire table is present or
not in a database if not it will create it first. After checking table
it will insert scrap data into MySQL database.